Move settings dialog to new lib
