+++ cspell.json
@@ -37,6 +37,7 @@
     "Pixmap",
     "Poschta",
     "Priddy",
+    "Regen",
     "runas",
     "Sbârnea",
     "Schoeneman",
