+++ src/gui/src/MainWindow.cpp
@@ -21,13 +21,13 @@
 #include "AboutDialog.h"
 #include "ActivationDialog.h"
 #include "ServerConfigDialog.h"
-#include "SettingsDialog.h"
 #include "gui/Logger.h"
 #include "gui/TrayIcon.h"
 #include "gui/VersionChecker.h"
 #include "gui/config/ConfigScopes.h"
 #include "gui/constants.h"
 #include "gui/core/CoreProcess.h"
+#include "gui/dialog/SettingsDialog.h"
 #include "gui/license/LicenseHandler.h"
 #include "gui/license/license_notices.h"
 #include "gui/messages.h"
