+++ src/lib/gui/dialog/SettingsDialogBase.ui
