+++ src/lib/gui/dialog/UpgradeDialog.h
@@ -1,7 +1,6 @@
 /*
  * synergy -- mouse and keyboard sharing utility
- * Copyright (C) 2012-2021 Symless Ltd.
- * Copyright (C) 2008 Volker Lanz (vl@fidra.de)
+ * Copyright (C) 2022 Symless Ltd.
  *
  * This package is free software; you can redistribute it and/or
  * modify it under the terms of the GNU General Public License
@@ -15,21 +14,13 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-#include "ComputerNameValidator.h"
 
-#include <QRegularExpression>
+#pragma once
 
-namespace validators {
+#include <QMessageBox>
 
-ComputerNameValidator::ComputerNameValidator(const QString &message)
-    : IStringValidator(message) {}
-
-bool ComputerNameValidator::validate(const QString &input) const {
-  const QRegularExpression re(
-      "^[\\w\\._-]{0,255}$", QRegularExpression::CaseInsensitiveOption);
-  auto match = re.match(input);
-  auto result = match.hasMatch();
-  return result;
-}
-
-} // namespace validators
+class UpgradeDialog : public QMessageBox {
+public:
+  explicit UpgradeDialog(QWidget *parent = nullptr);
+  void showDialog(const QString &text);
+};
