+++ src/lib/gui/validators/ComputerNameValidator.cpp
@@ -15,19 +15,21 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-#pragma once
+#include "ComputerNameValidator.h"
 
-#include "LineEditValidator.h"
-#include "gui/config/ScreenList.h"
-#include "validators/ValidationError.h"
+#include <QRegularExpression>
 
 namespace validators {
 
-class ScreenNameValidator : public LineEditValidator {
-public:
-  explicit ScreenNameValidator(
-      QLineEdit *lineEdit = nullptr, ValidationError *error = nullptr,
-      const ScreenList *pScreens = nullptr);
-};
+ComputerNameValidator::ComputerNameValidator(const QString &message)
+    : IStringValidator(message) {}
+
+bool ComputerNameValidator::validate(const QString &input) const {
+  const QRegularExpression re(
+      "^[\\w\\._-]{0,255}$", QRegularExpression::CaseInsensitiveOption);
+  auto match = re.match(input);
+  auto result = match.hasMatch();
+  return result;
+}
 
 } // namespace validators
