+++ src/lib/gui/validators/ComputerNameValidator.h
@@ -1,6 +1,6 @@
 /*
  * synergy -- mouse and keyboard sharing utility
- * Copyright (C) 2022 Symless Ltd.
+ * Copyright (C) 2021 Symless Ltd.
  *
  * This package is free software; you can redistribute it and/or
  * modify it under the terms of the GNU General Public License
@@ -14,11 +14,19 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-#pragma once
-#include <QMessageBox>
 
-class UpgradeDialog : public QMessageBox {
+#pragma once
+
+#include "IStringValidator.h"
+
+#include <QRegularExpression>
+
+namespace validators {
+
+class ComputerNameValidator : public IStringValidator {
 public:
-  explicit UpgradeDialog(QWidget *parent = nullptr);
-  void showDialog(const QString &text);
+  ComputerNameValidator(const QString &message);
+  bool validate(const QString &input) const override;
 };
+
+} // namespace validators
