+++ src/lib/gui/validators/EmptyStringValidator.h
@@ -1,7 +1,6 @@
 /*
  * synergy -- mouse and keyboard sharing utility
- * Copyright (C) 2012-2021 Symless Ltd.
- * Copyright (C) 2008 Volker Lanz (vl@fidra.de)
+ * Copyright (C) 2021 Symless Ltd.
  *
  * This package is free software; you can redistribute it and/or
  * modify it under the terms of the GNU General Public License
@@ -16,19 +15,16 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-#ifndef SPACESVALIDATOR_H
-#define SPACESVALIDATOR_H
+#pragma once
 
 #include "IStringValidator.h"
 
 namespace validators {
 
-class SpacesValidator : public IStringValidator {
+class EmptyStringValidator : public IStringValidator {
 public:
-  explicit SpacesValidator(const QString &message);
+  explicit EmptyStringValidator(const QString &message);
   bool validate(const QString &input) const override;
 };
 
 } // namespace validators
-
-#endif // SPACESVALIDATOR_H
