+++ src/lib/gui/validators/IStringValidator.h
@@ -1,7 +1,6 @@
 /*
  * synergy -- mouse and keyboard sharing utility
- * Copyright (C) 2012-2021 Symless Ltd.
- * Copyright (C) 2008 Volker Lanz (vl@fidra.de)
+ * Copyright (C) 2021 Symless Ltd.
  *
  * This package is free software; you can redistribute it and/or
  * modify it under the terms of the GNU General Public License
@@ -15,15 +14,23 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-#include "SpacesValidator.h"
+
+#pragma once
+
+#include <QString>
 
 namespace validators {
 
-SpacesValidator::SpacesValidator(const QString &message)
-    : IStringValidator(message) {}
+class IStringValidator {
+  QString m_Message;
 
-bool SpacesValidator::validate(const QString &input) const {
-  return !input.contains(' ');
-}
+public:
+  IStringValidator() = default;
+  explicit IStringValidator(const QString &message);
+  const QString &getMessage() const;
+
+  virtual bool validate(const QString &input) const = 0;
+  virtual ~IStringValidator() = default;
+};
 
 } // namespace validators
