+++ src/lib/gui/validators/LineEditValidator.cpp
