+++ src/lib/gui/validators/LineEditValidator.h
