+++ src/lib/gui/validators/ScreenDuplicationsValidator.cpp
