+++ src/lib/gui/validators/ScreenDuplicationsValidator.h
