+++ src/lib/gui/validators/SpacesValidator.cpp
@@ -1,7 +1,6 @@
 /*
  * synergy -- mouse and keyboard sharing utility
- * Copyright (C) 2012-2021 Symless Ltd.
- * Copyright (C) 2008 Volker Lanz (vl@fidra.de)
+ * Copyright (C) 2021 Symless Ltd.
  *
  * This package is free software; you can redistribute it and/or
  * modify it under the terms of the GNU General Public License
@@ -15,15 +14,16 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-#include "EmptyStringValidator.h"
+
+#include "SpacesValidator.h"
 
 namespace validators {
 
-EmptyStringValidator::EmptyStringValidator(const QString &message)
+SpacesValidator::SpacesValidator(const QString &message)
     : IStringValidator(message) {}
 
-bool EmptyStringValidator::validate(const QString &input) const {
-  return !input.isEmpty();
+bool SpacesValidator::validate(const QString &input) const {
+  return !input.contains(' ');
 }
 
 } // namespace validators
