+++ src/lib/gui/validators/SpacesValidator.h
@@ -1,7 +1,6 @@
 /*
  * synergy -- mouse and keyboard sharing utility
- * Copyright (C) 2012-2021 Symless Ltd.
- * Copyright (C) 2008 Volker Lanz (vl@fidra.de)
+ * Copyright (C) 2021 Symless Ltd.
  *
  * This package is free software; you can redistribute it and/or
  * modify it under the terms of the GNU General Public License
@@ -15,19 +14,17 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-#ifndef EMPTYSTRINGVALIDATOR_H
-#define EMPTYSTRINGVALIDATOR_H
+
+#pragma once
 
 #include "IStringValidator.h"
 
 namespace validators {
 
-class EmptyStringValidator : public IStringValidator {
+class SpacesValidator : public IStringValidator {
 public:
-  explicit EmptyStringValidator(const QString &message);
+  explicit SpacesValidator(const QString &message);
   bool validate(const QString &input) const override;
 };
 
 } // namespace validators
-
-#endif // EMPTYSTRINGVALIDATOR_H
