+++ src/lib/gui/validators/ValidationError.cpp
