+++ src/lib/gui/validators/ValidationError.h
