+++ src/lib/gui/widgets/ServerStateLabel.cpp
@@ -1,7 +1,6 @@
 /*
  * synergy -- mouse and keyboard sharing utility
- * Copyright (C) 2012-2021 Symless Ltd.
- * Copyright (C) 2008 Volker Lanz (vl@fidra.de)
+ * Copyright (C) 2021 Symless Ltd.
  *
  * This package is free software; you can redistribute it and/or
  * modify it under the terms of the GNU General Public License
@@ -22,7 +21,7 @@
 
 using namespace synergy::gui;
 
-namespace synergy_widgets {
+namespace synergy::gui::widgets {
 
 ServerStateLabel::ServerStateLabel(QWidget *parent) : QLabel(parent) {}
 
@@ -51,4 +50,4 @@ void ServerStateLabel::updateServerState(const QString &line) {
   }
 }
 
-} // namespace synergy_widgets
+} // namespace synergy::gui::widgets
